fixes overnight shift where end day is in the next month and not sunday
using the Calendar.DAY_OF_WEEK function to figure out if our shift is an overnight shift, we only took into consideration if the days are also at the end of the week cycle. this fix handles the case where end day is any other day. below is a list of days of the week and their number representations from the Calendar object
sun - 1
mon - 2
tues - 3
wed - 4
thurs - 5
fri - 6
sat - 7